requirements.txt
pymodbus==3.9.2
paho-mqtt==2.1.0
websocket-client==1.8.0